F4BY HAL

+++ libraries/AP_HAL_F4BY/NSHShellStream.cpp
@@ -122,7 +122,7 @@ int16_t NSHShellStream::read()
     return -1;
 }
 
-int16_t NSHShellStream::available()
+uint32_t NSHShellStream::available()
 {
     int ret = 0;
     if (ioctl(shell_stdin, FIONREAD, (unsigned long)&ret) == OK) {
@@ -131,7 +131,7 @@ int16_t NSHShellStream::available()
     return 0;
 }
 
-int16_t NSHShellStream::txspace()
+uint32_t NSHShellStream::txspace()
 {
     int ret = 0;
     if (ioctl(shell_stdout, FIONWRITE, (unsigned long)&ret) == OK) {
